+++ packages/dgt-id-proxy/lib/webid/webid-profile.handler.ts
@@ -1,32 +1,130 @@
+
+import * as path from 'path';
+import { readFile } from 'fs/promises';
 import { Handler } from '@digita-ai/handlersjs-core';
-import { Observable, of, throwError, from } from 'rxjs';
+import { Observable, of, throwError, from, zip } from 'rxjs';
 import { switchMap } from 'rxjs/operators';
 import { HttpHandlerResponse } from '@digita-ai/handlersjs-http';
 import { Writer, DataFactory } from 'n3';
+import { JWK, JWTPayload } from 'jose/types';
+import parseJwk from 'jose/jwk/parse';
+import SignJWT from 'jose/jwt/sign';
+import { v4 as uuid }  from 'uuid';
 
 const { literal } = DataFactory;
+
+export class PredicatePair {
+
+  constructor(public tokenKey: string, public predicate: string) {}
+
+}
+
+/**
+ * A {Handler} that handles a {HttpHandlerResponse} by checking if the webId given in the id_token
+ * has a WebId profile document and if not creates a new profile & .acl document for the webId.
+ */
 export class WebIdProfileHandler extends Handler<HttpHandlerResponse, HttpHandlerResponse> {
 
-  constructor(private predicates: { tokenKey: string; predicate: string }[]) {
+  /**
+   * Creates a {WebIdProfileHandler}.
+   *
+   * @param PredicatePair[]} predicates - the predicates that need to be set in the profile document
+   * @param {string} proxyWebId - the webId of the proxy server
+   * @param {string} pathToJwks - the path to a json file containing JWKs to sign the tokens.
+   */
+  constructor(
+    private predicates: PredicatePair[],
+    private proxyWebId: string,
+    private pathToJwks: string,
+    private proxyUrl: string,
+    private webIdPattern: string
+  ) {
 
     super();
 
-    if (!predicates) { throw new Error('Predicate list is required'); }
+    if (!predicates || predicates.length === 0) { throw new Error('Predicate list is required'); }
+
+    if (!proxyWebId) { throw new Error('WebId is required'); }
+
+    if (!pathToJwks) { throw new Error('Path to JWKS is required'); }
 
   }
 
+  /**
+   * Handles the response. Checks if an id_token and webId is present.
+   * Checks if a profile document already exists for the webId and if not creates a new one + acl document
+   *
+   * @param {HttpHandlerResponse} response
+   */
   handle(response: HttpHandlerResponse): Observable<HttpHandlerResponse> {
 
     if (!response) { return throwError(new Error('A response must be provided')); }
 
-    return from(this.getWebIdProfile(response.body.id_token.payload.webId)).pipe(
-      switchMap((data) => {
+    if (!response.body) { return throwError(new Error('A response body must be provided')); }
 
-        if (data.status !== 200) {
+    if (!response.body.id_token) { return throwError(new Error('An id token must be provided')); }
 
-          this.createWebIdProfile(
-            response.body.id_token.payload.webId,
-            this.generateProfileBody(this.predicates, response.body.id_token.payload.webId)
+    if (!response.body.id_token.payload.webid) { return throwError(new Error('A webId must be provided')); }
+
+    const id_token = response.body.id_token;
+    const webId = response.body.id_token.payload.webid;
+
+    // id_token.payload[].first_name = 'Alain';
+    // id_token.payload.family_name = 'Vandam';
+
+    // split the pattern to remove the ':'-prefixed claim
+    const splitPattern = this.webIdPattern.split(/:[a-zA-Z]+/g);
+
+    // create a new regex that matches anything that is put inplace of where the ':'-prefixed claim was
+    const regExp = new RegExp(splitPattern.join('?.*'));
+
+    if (!regExp.test(webId)) {
+
+      return throwError(new Error('The provided webId in the id_token must be similar to the webIdPattern'));
+
+    }
+
+    const jwt_payload: JWTPayload = {
+      iss: this.proxyUrl,
+      sub: id_token.payload.sub,
+      aud: 'solid',
+      jti: uuid(),
+      exp: Math.round((new Date()).getTime() / 1000) + 7200,
+      iat: Math.round((new Date()).getTime() / 1000),
+      webid: this.proxyWebId,
+
+    };
+
+    return from(fetch(webId, { method: 'HEAD', headers: { Accept: 'text/turtle' } })).pipe(
+      switchMap((resp) => {
+
+        if (resp.status !== 200) {
+
+          return this.signJwtPayload(jwt_payload, 'at+jwt').pipe(
+            switchMap((token) => {
+
+              const proxyUrl = new URL(this.proxyWebId);
+              const userUrl = new URL(webId);
+              // example: http://localhost:3002/clientapp/tonypaillard/profile/card#me
+              const proxyURI = proxyUrl.origin + userUrl.pathname;
+              // example: http://localhost:3002/clientapp/tonypaillard/profile/card.acl
+              const aclURI = proxyUrl.origin + userUrl.pathname + '.acl';
+              const profileResp = from(fetch(proxyURI, { method: 'PUT', headers: { Accept: 'text/turtle',  Authorization:  'Bearer ' + token, 'Content-Type': 'text/turtle' }, body: this.generateProfileDocument(id_token) }));
+              const aclResp = from(fetch(aclURI, { method: 'PUT', headers: { Accept: 'text/turtle',  Authorization:  'Bearer ' + token, 'Content-Type': 'text/turtle' }, body: this.generateAclDocument(id_token.payload.webid) }));
+
+              return zip(profileResp, aclResp).pipe(
+                switchMap(([ profile, acl ]) => {
+
+                  if (profile.status !== 200 && profile.status !== 201 && profile.status !== 205) return throwError(new Error('Failed to create a profile document'));
+
+                  if (acl.status !== 200 && acl.status !== 201 && profile.status !== 205) return throwError(new Error('Failed to create Acl document'));
+
+                  return of(response);
+
+                })
+              );
+
+            }),
           );
 
         }
@@ -38,57 +136,97 @@ export class WebIdProfileHandler extends Handler<HttpHandlerResponse, HttpHandle
 
   }
 
+  /**
+   * Specifies that if the response is defined this handler can handle the response.
+   *
+   * @param {HttpHandlerResponse} response
+   */
   canHandle(response: HttpHandlerResponse): Observable<boolean> {
 
     return response ? of(true) : of(false);
 
   }
 
-  getWebIdProfile = async (webId: string): Promise<Response>  => {
+  /**
+   * Generates an acl document based on the webId provided.
+   *
+   * @param {string} webId - the webId provided in the id_token
+   */
+  private generateAclDocument(webId: string): string {
 
-    const data = await fetch(webId, {
-      method: 'HEAD',
-      headers: {
-        Accept: 'text/turtle',
-      },
-    });
+    return `# ACL resource for the WebID profile document
+    @prefix acl: <http://www.w3.org/ns/auth/acl#>.
+    @prefix foaf: <http://xmlns.com/foaf/0.1/>.
 
-    return data;
+    # The WebID profile is readable by the public.
+    # This is required for discovery and verification,
+    # e.g. when checking identity providers.
+    <#public>
+        a acl:Authorization;
+        acl:agentClass foaf:Agent;
+        acl:accessTo <./card>;
+        acl:mode acl:Read.
 
-  };
+    # The owner has full access to the entire
+    # profile directory.
+    <#owner>
+        a acl:Authorization;
+        acl:agent <${webId}>;
+        acl:accessTo <./card>;
+        acl:mode acl:Read, acl:Write, acl:Control.`;
 
-  createWebIdProfile = async (webId: string, body: string): Promise<void>  => {
+  }
 
-    await fetch(webId, {
-      method: 'PUT',
-      headers: {
-        Accept: 'text/turtle',
-      },
-      body,
-    });
+  /**
+   * Generates a profile document based on the predicates provided in the constructor and the id_token.
+   *
+   * @param {{ header: any; payload: any }} id_token - the id_token provided in the response.
+   *
+   */
+  private generateProfileDocument(id_token:  { header: any; payload: any }): string {
 
-  };
-
-  generateProfileBody = (predicates: { tokenKey: string; predicate: string }[], webId: string): string => {
+    const webId = id_token.payload.webid;
 
     const writer = new Writer({ prefixes: { foaf: 'http://xmlns.com/foaf/0.1/', solid: 'http://www.w3.org/ns/solid/terms#' } });
     let body = '';
 
-    predicates.forEach((keyPredicatePair) => {
+    this.predicates.forEach((keyPredicatePair) => {
 
       writer.addQuad(
         DataFactory.namedNode(webId),
         DataFactory.namedNode(keyPredicatePair.tokenKey),
-        literal(keyPredicatePair.predicate),
+        literal(id_token.payload[keyPredicatePair.tokenKey]),
       );
 
     });
+
+    writer.addQuad(
+      DataFactory.namedNode(webId),
+      DataFactory.namedNode('solid:oidcIssuer'),
+      DataFactory.namedNode(this.proxyUrl)
+    );
 
     writer.end((err, result) => body = result);
 
     return body;
 
-  };
+  }
+
+  private getSigningKit = () => from(readFile(
+    path.isAbsolute(this.pathToJwks) ? this.pathToJwks : path.join(process.cwd(), this.pathToJwks)
+  )).pipe(
+    switchMap<Buffer, JWK>((keyFile) => of(JSON.parse(keyFile.toString()).keys[0])),
+    switchMap((jwk) => zip(of(jwk.alg), of(jwk.kid), from(parseJwk(jwk)))),
+  );
+
+  private signJwtPayload = (jwtPayload: JWTPayload, typ: string) => zip(of(jwtPayload), this.getSigningKit()).pipe(
+    switchMap(([ payload, [ alg, kid, key ] ]) => from(
+      new SignJWT(payload)
+        .setProtectedHeader({ alg, kid, typ })
+        .setJti(uuid())
+        .sign(key),
+    )),
+  );
 
 }