+++ packages/dgt-id-proxy/CHANGELOG.md
@@ -3,6 +3,46 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+### [0.8.7](https://github.com/digita-ai/dgt-id-broker/compare/v0.8.6...v0.8.7) (2021-10-26)
+
+**Note:** Version bump only for package @digita-ai/dgt-id-proxy
+
+
+
+
+
+### [0.8.6](https://github.com/digita-ai/dgt-id-broker/compare/v0.8.5...v0.8.6) (2021-10-26)
+
+**Note:** Version bump only for package @digita-ai/dgt-id-proxy
+
+
+
+
+
+### [0.8.5](https://github.com/digita-ai/dgt-id-broker/compare/v0.8.4...v0.8.5) (2021-10-25)
+
+**Note:** Version bump only for package @digita-ai/dgt-id-proxy
+
+
+
+
+
+### [0.8.4](https://github.com/digita-ai/dgt-id-broker/compare/v0.8.3...v0.8.4) (2021-10-25)
+
+**Note:** Version bump only for package @digita-ai/dgt-id-proxy
+
+
+
+
+
+### [0.8.3](https://github.com/digita-ai/dgt-id-broker/compare/v0.8.2...v0.8.3) (2021-10-25)
+
+**Note:** Version bump only for package @digita-ai/dgt-id-proxy
+
+
+
+
+
 ### [0.8.2](https://github.com/digita-ai/dgt-id-broker/compare/v0.8.1...v0.8.2) (2021-10-22)